backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-cloud-logging
google-cloud-aiplatform
python-dotenv

python-multipart
